+++ envoy-control-tests/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/config/service/EchoContainer.kt
@@ -6,7 +6,7 @@ import pl.allegro.tech.servicemesh.envoycontrol.config.testcontainers.GenericCon
 import java.util.UUID
 import java.util.Locale
 
-class EchoContainer : GenericContainer<EchoContainer>("hashicorp/http-echo:latest"), ServiceContainer {
+class EchoContainer : GenericContainer<EchoContainer>("jxlwqq/http-echo"), ServiceContainer {
 
     val response = UUID.randomUUID().toString()
 
@@ -14,7 +14,7 @@ class EchoContainer : GenericContainer<EchoContainer>("hashicorp/http-echo:lates
         super.configure()
         withExposedPorts(PORT)
         withNetwork(Network.SHARED)
-        withCommand(String.format(Locale.getDefault(), "-text=%s", response))
+        withCommand(String.format(Locale.getDefault(), "--text=%s --addr=:%s", response, PORT))
         waitingFor(Wait.forHttp("/").forStatusCode(200))
     }